+++ sr_robot_launch/test/test_sr_ur_unlock.test
@@ -1,6 +1,8 @@
 <launch>
-  <param name="ra_sr_ur_robot_hw/headless_mode" value="true"/>
-  <param name="la_sr_ur_robot_hw/headless_mode" value="true"/>
+  <!--param name="ra_sr_ur_robot_hw/headless_mode" value="true"/>
+  <param name="la_sr_ur_robot_hw/headless_mode" value="true"/-->
 
-  <test test-name="test_sr_ur_unlock" pkg="sr_robot_launch" type="test_sr_ur_unlock.py" time-limit="700.0"/>
+  <test test-name="test_sr_ur_unlock_left" pkg="sr_robot_launch" type="test_sr_ur_unlock_left.py" time-limit="700.0"/>
+  <test test-name="test_sr_ur_unlock_right" pkg="sr_robot_launch" type="test_sr_ur_unlock_right.py" time-limit="700.0"/>
+  <test test-name="test_sr_ur_unlock_bimanual" pkg="sr_robot_launch" type="test_sr_ur_unlock_bimanual.py" time-limit="700.0"/>
 </launch>
